fix: recordHPをdouble型に変更

+++ grenze/lib/user_data.dart
@@ -255,8 +255,8 @@ class UserDataProvider with ChangeNotifier {
     updateMinMaxSpots();
     imgUrl = responseBody["url"];
     avatarName = responseBody["avatarName"];
-    recordHighHP = responseBody["recordHighHP"];
-    recordLowHP = responseBody["recordLowHP"];
+    recordHighHP = responseBody["recordHighHP"].toDouble();
+    recordLowHP = responseBody["recordLowHP"].toDouble();
     activeLimitTime = responseBody["activeLimitTime"];
     maxSleepDuration = responseBody["maxSleepDuration"];
     maxDayHP = responseBody["maxDayHP"];
